requirements.txt
streamlit==1.28.0
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
faiss-cpu==1.7.4
opencv-python==4.8.1.78
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
plotly>=5.15.0
tqdm>=4.65.0
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0